// src/types/car.type.ts
export type TCar = {
    _id: string;
    name: string;
    model: string;
    photo: string;
    description: string;
    color: string;
    features: string[];
    isElectric: boolean;
    pricePerHour: number;
    status: string;
    isDeleted: boolean;
    createdAt: string;
    updatedAt: string;
    type?: string;
    __v: number;
  };
  